GCC now defaults to -fno-common. Use the common module variables.c for global variables.
